Configuration module refactor
